Wizert moves error
im so frustrated, i can make wizert change his location in the array but this requires a long if else statement. i wish there was an easier way to do this. I cant seem to just make it so wizert can have the option to move without a long hard coding